Move core KLGraphics functions out of __init__
